build: fetch favicons for tidbits also
