+++ packages/apps/web/package.json
@@ -7,8 +7,8 @@
     "pkg:nuke": "pnpm run pkg:nuke:artifacts && rimraf node_modules",
     "pkg:nuke:artifacts": "rimraf dist generated/favicons-for-websites.json *.tsbuildinfo",
     "dev": "next dev",
-    "pkg:build": "pnpm run pkg:nuke:artifacts && pnpm run build:fetch-favicons-for-posts && pnpm run build:next && pnpm run build:sitemap",
-    "build:fetch-favicons-for-posts": "ts-node --esm --project ./tsconfig.project.json ./scripts/fetch-favicons-for-posts.ts",
+    "pkg:build": "pnpm run pkg:nuke:artifacts && pnpm run build:fetch-favicons-for-articles && pnpm run build:next && pnpm run build:sitemap",
+    "build:fetch-favicons-for-articles": "ts-node --esm --project ./tsconfig.project.json ./scripts/fetch-favicons-for-articles.ts",
     "build:next": "next build",
     "build:sitemap": "next-sitemap --config next-sitemap.cjs",
     "export": "next export",
@@ -23,7 +23,7 @@
   },
   "nx": {
     "targets": {
-      "build": {
+      "pkg:build": {
         "outputs": [
           "{projectRoot}/dist/**/*",
           "{projectRoot}/generated/**/*"
